后台用户模块dao层 service层 以及service层测试

+++ src/main/java/com/example/pethospitalbackend/service/VideoService.java
@@ -41,9 +41,9 @@ public class VideoService {
     }
     // 修改数据库
     TreatmentVideo video = new TreatmentVideo();
-    video.setCaseId((long) 1);
+    video.setCaseId(1L);
     video.setUrl(url);
-    video.setSortNum((long) 4);
+    video.setSortNum(4L);
 
     boolean result = treatmentVideoDao.insertVideo(video) > 0;
